Introduce IpczDriver layer
This abstracts away I/O details, leaving them up to the embedding
application. Introduces kSingleProcessReferenceDriver and
kMultiprocessReferenceDriver as reference implementations to use in ipcz
tests.

The single-process implementation passes messages directly into peer
endpoints, which effectively results in synchronous behavior for all
cross-node ipcz operations.

The multiprocess implementation is a thin wrapper around os::Channel,
using a per-Channel background thread to monitor for incoming messages.

In addition to this change, OpenRemotePortals and AcceptRemotePortals
are replaced with a unified ConnectNode API which allows for multiple
initial portals and uses flags to hint at the intended relationship
between the caller and the remote node.