semi-working direct linkage of colocated peers
when a portal is sent to the node which houses its peer, we want to
establish a direct in-memory link between them in order to circumvent
the routing layer.

this introduces a test case which exercising bouncing a portal back and
forth between nodes.

a second test case is introduced to move both portals from a pair
simultaneously across multiple separate nodes, bouncing each to and from
the broker at each step. this case is still broken with an async driver,
because we don't anticipate a member of a split local pair having a
predecessor link.