only defer trap events when trap is armed
this is in contrast with the previous logic which could allow a trap to
defer an arbitrary number of events in response to a single state
change. even though any of those events would disarm the trap when
executed, the execution could in turn re-arm the same trap and allow
other already-deferred events to execute immediately thereafter. that's
not what we want.