partial half-proxy decay impl
implements a small piece of half-proxy decay. spcefically this allows
serialized routers to half-proxy immediately if they have a peer that is
not a half-proxy. this results in its successor immediately sending a
BypassProxy to the half-proxy's peer.

BypassProxy is not currently handled though, and Node discovery is not
(re-)implemented yet.

this also fixes several issues with local pair splitting which were
obscured by another bug that essentially rendered it dead code before.
now local pairs are actually split properly when possible.