non-polarized Routers
Routers no longer have an objective "side" to which they belong.
Instead, the two ends of each individual link have an arbitrary side A
and side B agreed upon by convention at construction time, and a route
in general has two relative sides which can be referred to as
RouteSide::kSame or RouteSide::kOther, from an individual Router's
perspective.

This model is equally capable of supporting the existing implementation,
but it also makes a new MergePoratls API tractable to implement.