introduce 3- and 4-mutex lockers
These are needed for the relatively rare operation of bridged route
decay so that all necessary locks can be acquired in a globally
consistent order.

Factors out a common internal MultiMutexLock template helper to sort an
array of Mutex* and lock them in address order.

Using a template helper directly is unwieldy because we can't convey a
variadic list of mutexes to clang's exclusive_lock_function() attribute