implement broker inheritance
if a non-broker connects to another non-broker, one of them must already
have a broker connection or must assume responsibility for establishing
one. that non-broker can specify IPCZ_CONNECT_NODE_SHARE_BROKER and the
other non-broker can specify IPCZ_CONNECT_NODE_INHERIT_BROKER. this
allows the first non-broker to arrange for the second one to get both a
link to the first one's broker, as well as a link (with initial portals)
to the first one itself.