Improve link state synchronization
This changes RouterLinkState to represent its status as a bitfield. Bits
can be set to indicate that either side of the link is stable, waiting,
and/or locked.

A link can only be locked by one side at a time and only while both
sides are stable. If one side of a link attempts to lock while the other
side is not stable, the first side instead sets its waiting bit.

When a router on the central link is flushed in a state where bypass may
be supported it will attempt to lock the link and initiate its own
bypass. If that fails it will test-and-reset the other side's waiting
bit, pinging the other side to flush if the bit was set.

This also simplifies route closure propagation by imposing two
preconditions: only a terminal router on the central link can propagate
closure outward, and the central link must be locked before doing so.
This ensures that closure propagation traverses a stable route.