Fix ordering bug on NodeLinkMemory buffer sharing
When a node allocates a buffer for NodeLinkMemory, it sends a handle to
the buffer to the other end of the NodeLink ASAP.

It also has a collection of callbacks which are accumulated to execute
as soon as the buffer is available (e.g. to establish latent
RouterLinkState address mappings).

Those callbacks can send communications across the NodeLink which assume
the other side already has the buffer available to it.

Before this change, the callbacks were run *before* sharing the buffer
with the remote node. This change shares the buffer first.

Fixes a rare deadlock in RemotePortalTest.TransferManyLocalPortals/2.