Introduce notion of addressability to fragments
Some non-null fragments are not addressable until a local NodeLinkMemory
has access to their underlying buffer. This allows for the existence of
a Fragment and FragmentRef which is non-null but which still has no
mapped address to reference.

Maybe a little sloppy, maybe there should be a separate type for this,
or something like a FutureFragmentRef for async resolution.

In any case, for now the only use of fragments (RemoteRouterLink's
RouterLinkState) is updated to gracefully handle non-addressable
fragments by waiting for the identified buffer to become available.

This allows for arbitrary latency between a link being established with
a specific fragment ref, and the owning node having access to the
corresponding shared buffer; which is critical for Windows support where
handles (like shared memory handles) must be relayed by a broker
process when passing between two non-brokers.