Consolidation and documentation
Most of the code in src/os/ is not used by the ipcz implementation, but
is instead there to support testing and reference drivers.

This renames src/drivers/ to src/reference_drivers/ and moves Channel,
Event, and Memory implementations there.

Also adds brief documentation to src/reference_drivers/ and
src/standalone/ to clarify their use and the fact that they are not part
of the ipcz implementation itself.