Implement message relay through brokers
Windows handles need to be manually duplicated between processes, and
arbitrary processes cannot in general be trusted to do this.

So for messages which might go from one non-broker to another while
carrying an OS handle, this change forces them to be relayed through
a broker instead of being transmitted directly.

Relaying behavior currently affects only AcceptParcel and
AddFragmentAllocatorBuffer messages, and is only enabled on Windows.