The great OS handle purge
The concept of an "OS handle" is a sloppy abstraction for ipcz to
expose, leaking other platform-specific details into the implementation.
This risks obsolence in light of new platforms or new platform features,
and it risks ossification of subtle details that might render the
library more difficult to use or maintain within certain environments.

This removes IpczOSHandle and IpczOSProcessHandle from the ipcz API.
ConnectNode no longer takes an IpczOSProcessHandle, and Put/Get etc APIs
no longer accept collections of OS handles.

Instead drivers are expected to facilitate transfer of such objects on
their own, and the driver API is modified to support this. To wit,
driver Serialize() no longer translates a driver handle to a collection
of bytes and OS handles: instead it produces a collection of bytes and
other driver handles, and the output handles must be transmissible as-is
by the driver. For example, on POSIX systems a driver might wrap a file
descriptor as a driver object and return a driver handle to it as as a
serialization artifact from a more complex driver object. The driver
can trivially transmit the descriptor out-of-band alongside message
data if using a UNIX socket as the underlying transport.

Deserialize(), Transmit(), and IpczTransportAcivityHandler have also
been updated accordingly to pass these transmissible driver handles
between ipcz and its driver.

With drivers able to effectively encapsulate platform handle transfer on
their own, applications can rely on object boxing to get platform
handles in and out of portals.