Fix reference drivers on Windows
After the purging of OS handles from ipcz, the Windows reference
drivers needed to be adapted to support HANDLE transfer. This does
that.

Also tweaks the driver's Serialize and Deserialize APIs to provide
the transmitting transport as context, and to allow a driver to
indicate that an object is serializable but not transmissible over
a specific transport. This provides the basis for ipcz relaying
messages through a broker, which will in turn support practical
multiprocess drivers on Windows.