Preparation for test coverage of relayed objects
This updates the driver API such that CreateTransports() now takes a
reference to two existing transports as context for the new transport
being created -- more specifically, the driver can assume that the new
transport will be used to directly connect the two input transports'
respective remote nodes.

The multiprocess reference driver transport is also updated to support
awareness of the broker/non-broker relationship, and test code is
updated to plumb relevant details accordingly.