Lift FragmentAllocator above block allocation
NodeLinkMemory should be able to conveniently allocate larger chunks of
shared memory with non-uniform size, e.g. for streamed data transfer.
BlockAllocator is not well suited for this but works great for the
common small fragments needed by RouterLinkState and most application
messages.

This effectively renames the existing FragmentAllocator to
BlockAllocatorPool, then rebuilds FragmentAllocator with a slightly more
generic interface which, for now, simply wraps ownership of multiple
BlockAllocatorPools.

This is a precursor to supporting larger allocations through
FragmentAllocator.